+++ src/pages/farmer/EditFarmer.tsx
@@ -79,15 +79,11 @@ const EditFarmer = () => {
   const navigate = useNavigate();
 
   const farmerId = queryString[1];
-  const [data, setData] = useState<GetFarmerEntity>(
-    GetFarmerEntity_INIT
-  );
-  const [address, setAddress] = useState<AddressEntity>(
-    AddressEntity_INIT
-  );
-  const [farmerPlotList, setFarmerPlotList] = useState<
-    FarmerPlotEntity[]
-  >([FarmerPlotEntity_INIT]);
+  const [data, setData] = useState<GetFarmerEntity>(GetFarmerEntity_INIT);
+  const [address, setAddress] = useState<AddressEntity>(AddressEntity_INIT);
+  const [farmerPlotList, setFarmerPlotList] = useState<FarmerPlotEntity[]>([
+    FarmerPlotEntity_INIT,
+  ]);
   const [showAddModal, setShowAddModal] = useState(false);
   const [showEditModal, setShowEditModal] = useState(false);
   const [editIndex, setEditIndex] = useState(0);
@@ -98,13 +94,13 @@ const EditFarmer = () => {
   const [district, setDistrict] = useState<DistrictEntity[]>([
     DistrictEntity_INIT,
   ]);
-  const [subdistrict, setSubdistrict] = useState<SubdistrictEntity[]>(
-    [SubdistrictEntity_INIT]
+  const [subdistrict, setSubdistrict] = useState<SubdistrictEntity[]>([
+    SubdistrictEntity_INIT,
+  ]);
+  const [saveBtnDisable, setBtnSaveDisable] = useState<boolean>(false);
+  const [editFarmerPlot, setEditFarmerPlot] = useState<FarmerPlotEntity>(
+    FarmerPlotEntity_INIT
   );
-  const [saveBtnDisable, setBtnSaveDisable] =
-    useState<boolean>(false);
-  const [editFarmerPlot, setEditFarmerPlot] =
-    useState<FarmerPlotEntity>(FarmerPlotEntity_INIT);
 
   const [imgProfile, setImgProfile] = useState<any>();
   const [imgIdCard, setImgIdCard] = useState<any>();
@@ -123,12 +119,8 @@ const EditFarmer = () => {
 
       setAddress(res.address);
       setFarmerPlotList(res.farmerPlot);
-      let getPathPro = res.file.filter(
-        (x) => x.category === "PROFILE_IMAGE"
-      );
-      let getPathCard = res.file.filter(
-        (x) => x.category === "ID_CARD_IMAGE"
-      );
+      let getPathPro = res.file.filter((x) => x.category === "PROFILE_IMAGE");
+      let getPathCard = res.file.filter((x) => x.category === "ID_CARD_IMAGE");
       let imgList: {
         path: string;
         category: string;
@@ -143,21 +135,21 @@ const EditFarmer = () => {
       let i = 0;
       for (i; imgList.length > i; i++) {
         i === 0 &&
-          UploadImageDatasouce.getImage(
-            imgList[i].path.toString()
-          ).then((resImg) => {
-            if (resImg.url) {
-              imgList[0].category === "PROFILE_IMAGE"
-                ? setImgProfile(resImg.url)
-                : setImgIdCard(resImg.url);
+          UploadImageDatasouce.getImage(imgList[i].path.toString()).then(
+            (resImg) => {
+              if (resImg.url) {
+                imgList[0].category === "PROFILE_IMAGE"
+                  ? setImgProfile(resImg.url)
+                  : setImgIdCard(resImg.url);
+              }
             }
-          });
+          );
         i === 1 &&
-          UploadImageDatasouce.getImage(
-            imgList[i].path.toString()
-          ).then((resImg) => {
-            resImg?.url && setImgIdCard(resImg.url);
-          });
+          UploadImageDatasouce.getImage(imgList[i].path.toString()).then(
+            (resImg) => {
+              resImg?.url && setImgIdCard(resImg.url);
+            }
+          );
       }
     });
   };
@@ -172,18 +164,14 @@ const EditFarmer = () => {
       setProvince(res);
     });
     if (address.provinceId) {
-      LocationDatasource.getDistrict(address.provinceId).then(
-        (res) => {
-          setDistrict(res);
-        }
-      );
+      LocationDatasource.getDistrict(address.provinceId).then((res) => {
+        setDistrict(res);
+      });
     }
     if (address.districtId) {
-      LocationDatasource.getSubdistrict(address.districtId).then(
-        (res) => {
-          setSubdistrict(res);
-        }
-      );
+      LocationDatasource.getSubdistrict(address.districtId).then((res) => {
+        setSubdistrict(res);
+      });
     }
   }, [address.provinceId, address.districtId]);
 
@@ -237,16 +225,14 @@ const EditFarmer = () => {
     checkValidateAddr(c.toJS());
   };
 
-  const handleOnChangeAddress = (
-    e: React.ChangeEvent<HTMLTextAreaElement>
-  ) => {
+  const handleOnChangeAddress = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     const d = Map(address).set("address1", e.target.value);
     setAddress(d.toJS());
     checkValidateAddr(d.toJS());
   };
 
   const handleChangeFarmerstatus = (e: any) => {
-    if (e.target.value != "REJECTED") {
+    if (e.target.value != "INATIVE" || e.target.value != "REJECTED") {
       data.reason = "";
     }
     const m = Map(data).set("status", e.target.value);
@@ -255,9 +241,7 @@ const EditFarmer = () => {
     checkValidateReason(m.toJS());
   };
 
-  const handleOnChangeReason = (
-    e: React.ChangeEvent<HTMLTextAreaElement>
-  ) => {
+  const handleOnChangeReason = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     const d = Map(data).set("reason", e.target.value);
     setData(d.toJS());
     checkValidateAddr(d.toJS());
@@ -355,9 +339,7 @@ const EditFarmer = () => {
   };
 
   const removeImg = () => {
-    const getImg = data.file.filter(
-      (x) => x.category == "PROFILE_IMAGE"
-    )[0];
+    const getImg = data.file.filter((x) => x.category == "PROFILE_IMAGE")[0];
     if (getImg !== undefined) {
       UploadImageDatasouce.deleteImage(getImg.id, getImg.path).then(
         (res) => {}
@@ -414,9 +396,7 @@ const EditFarmer = () => {
     imgWindow?.document.write(image.outerHTML);
   };
   const removeImgIdCard = () => {
-    const getImg = data.file.filter(
-      (x) => x.category == "ID_CARD_IMAGE"
-    )[0];
+    const getImg = data.file.filter((x) => x.category == "ID_CARD_IMAGE")[0];
     if (getImg != undefined) {
       UploadImageDatasouce.deleteImage(getImg.id, getImg.path).then(
         (res) => {}
@@ -487,19 +467,17 @@ const EditFarmer = () => {
     };
     delete payload.farmerPlot;
     await FarmerDatasource.updateFarmer(payload).then((res) => {
+      console.log(payload);
+      console.log(res);
       if (res !== undefined) {
         let i = 0;
         for (i; 2 > i; i++) {
           i === 0 &&
             createImgProfile.file !== "" &&
-            UploadImageDatasouce.uploadImage(createImgProfile).then(
-              res
-            );
+            UploadImageDatasouce.uploadImage(createImgProfile).then(res);
           i === 1 &&
             createImgIdCard.file !== "" &&
-            UploadImageDatasouce.uploadImage(createImgIdCard).then(
-              res
-            );
+            UploadImageDatasouce.uploadImage(createImgIdCard).then(res);
         }
         Swal.fire({
           title: "บันทึกสำเร็จ",
@@ -532,7 +510,8 @@ const EditFarmer = () => {
                   backgroundImage: `url(${
                     imgProfile === undefined ? img_empty : imgProfile
                   })`,
-                }}>
+                }}
+              >
                 <input
                   key={imgProfile}
                   type="file"
@@ -549,7 +528,8 @@ const EditFarmer = () => {
                       style={{
                         cursor: "pointer",
                         borderRadius: "5px",
-                      }}>
+                      }}
+                    >
                       View
                     </Tag>
                     <Tag
@@ -558,7 +538,8 @@ const EditFarmer = () => {
                       style={{
                         cursor: "pointer",
                         borderRadius: "5px",
-                      }}>
+                      }}
+                    >
                       Remove
                     </Tag>
                   </>
@@ -576,11 +557,8 @@ const EditFarmer = () => {
                     padding: "4px 12px",
                   }}
                   disabled
-                  value={`${moment(data.createdAt).format(
-                    "DD/MM/YYYY"
-                  )} ${
-                    data.createBy === null ||
-                    data.createBy === undefined
+                  value={`${moment(data.createdAt).format("DD/MM/YYYY")} ${
+                    data.createBy === null || data.createBy === undefined
                       ? "ลงทะเบียนโดยเกษตรกร"
                       : `(${data.createBy})`
                   }`}
@@ -600,7 +578,8 @@ const EditFarmer = () => {
                     required: true,
                     message: "กรุณากรอกชื่อ!",
                   },
-                ]}>
+                ]}
+              >
                 <Input
                   placeholder="กรอกชื่อ"
                   defaultValue={data.firstname}
@@ -620,7 +599,8 @@ const EditFarmer = () => {
                     required: true,
                     message: "กรุณากรอกนามสกุล!",
                   },
-                ]}>
+                ]}
+              >
                 <Input
                   placeholder="กรอกนามสกุล"
                   defaultValue={data.lastname}
@@ -650,7 +630,8 @@ const EditFarmer = () => {
                     min: 10,
                     message: "กรุณากรอกเบอร์โทรให้ครบ 10 หลัก!",
                   },
-                ]}>
+                ]}
+              >
                 <Input
                   placeholder="กรอกเบอร์โทร"
                   defaultValue={data.telephoneNo}
@@ -671,7 +652,8 @@ const EditFarmer = () => {
                     required: true,
                     message: "กรุณากรอกวันเดือนปีเกิด",
                   },
-                ]}>
+                ]}
+              >
                 <DatePicker
                   placeholder="กรอกวันเดือนปีเกิด"
                   format={dateFormat}
@@ -693,14 +675,14 @@ const EditFarmer = () => {
                 rules={[
                   {
                     min: 13,
-                    message:
-                      "กรุณากรอกรหัสบัตรประชาชนให้ครบ 13 หลัก!",
+                    message: "กรุณากรอกรหัสบัตรประชาชนให้ครบ 13 หลัก!",
                   },
                   {
                     pattern: new RegExp(/^[0-9\b]+$/),
                     message: "กรุณากรอกรหัสบัตรประชาชนให้ถูกต้อง!",
                   },
-                ]}>
+                ]}
+              >
                 <Input
                   placeholder="กรอกรหัสบัตรประชาชน"
                   defaultValue={data.idNo}
@@ -720,9 +702,9 @@ const EditFarmer = () => {
                   className="hiddenFileInput"
                   style={{
                     backgroundImage: `url(${imgIdCard})`,
-                    display:
-                      imgIdCard != undefined ? "block" : "none",
-                  }}></div>
+                    display: imgIdCard != undefined ? "block" : "none",
+                  }}
+                ></div>
               </div>
               <div className="text-left ps-4">
                 {imgIdCard != undefined && (
@@ -733,7 +715,8 @@ const EditFarmer = () => {
                       style={{
                         cursor: "pointer",
                         borderRadius: "5px",
-                      }}>
+                      }}
+                    >
                       View
                     </Tag>
                     <Tag
@@ -742,7 +725,8 @@ const EditFarmer = () => {
                       style={{
                         cursor: "pointer",
                         borderRadius: "5px",
-                      }}>
+                      }}
+                    >
                       Remove
                     </Tag>
                   </>
@@ -753,7 +737,8 @@ const EditFarmer = () => {
                 style={{
                   backgroundImage: `url(${bth_img_empty})`,
                   display: imgIdCard == undefined ? "block" : "none",
-                }}>
+                }}
+              >
                 <input
                   key={imgIdCard}
                   type="file"
@@ -783,11 +768,10 @@ const EditFarmer = () => {
                       .toLowerCase()
                       .localeCompare(optionB.children.toLowerCase())
                   }
-                  onChange={handleOnChangeProvince}>
+                  onChange={handleOnChangeProvince}
+                >
                   {province?.map((item) => (
-                    <Option value={item.provinceId}>
-                      {item.provinceName}
-                    </Option>
+                    <Option value={item.provinceId}>{item.provinceName}</Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -811,11 +795,10 @@ const EditFarmer = () => {
                       .toLowerCase()
                       .localeCompare(optionB.children.toLowerCase())
                   }
-                  onChange={handleOnChangeDistrict}>
+                  onChange={handleOnChangeDistrict}
+                >
                   {district?.map((item) => (
-                    <Option value={item.districtId}>
-                      {item.districtName}
-                    </Option>
+                    <Option value={item.districtId}>{item.districtName}</Option>
                   ))}
                 </Select>
               </Form.Item>
@@ -841,7 +824,8 @@ const EditFarmer = () => {
                       .toLowerCase()
                       .localeCompare(optionB.children.toLowerCase())
                   }
-                  onChange={handleOnChangeSubdistrict}>
+                  onChange={handleOnChangeSubdistrict}
+                >
                   {subdistrict?.map((item) => (
                     <Option value={item.subdistrictId}>
                       {item.subdistrictName}
@@ -876,7 +860,8 @@ const EditFarmer = () => {
                     required: true,
                     message: "กรุณากรอกที่อยู่บ้าน!",
                   },
-                ]}>
+                ]}
+              >
                 <TextArea
                   className="col-lg-12"
                   rows={5}
@@ -896,39 +881,51 @@ const EditFarmer = () => {
               <br />
               <Radio.Group
                 defaultValue={data.status}
-                onChange={handleChangeFarmerstatus}>
+                onChange={handleChangeFarmerstatus}
+              >
                 <Space direction="vertical">
-                  {FARMER_STATUS_SEARCH.map((item) => (
-                    <Radio value={item.value}>{item.name}</Radio>
+                  {FARMER_STATUS_SEARCH.map((item, index) => (
+                    <Radio value={item.value}>
+                      {item.name}
+                      {(data.status == "REJECTED" && index == 2) ||
+                      (data.status == "INACTIVE" && index == 3) ? (
+                        <div>
+                          <div className="form-group">
+                            <label></label>
+                            <br />
+                            <Form.Item>
+                              <TextArea
+                                className="col-lg-12"
+                                rows={3}
+                                placeholder="กรอกเหตุผล/เหตุหมายเพิ่มเติม"
+                                autoComplete="off"
+                                onChange={handleOnChangeReason}
+                                defaultValue={data.reason}
+                              />
+                            </Form.Item>
+                          </div>
+                        </div>
+                      ) : null}
+                    </Radio>
                   ))}
                 </Space>
               </Radio.Group>
             </div>
           </div>
-          {data.status == "REJECTED" && (
-            <div>
-              <div className="form-group">
-                <label></label>
-                <br />
-                <Form.Item name="reason"
-                  rules={[
-                    {
-                      required: data.status === "REJECTED",
-                      message: "กรุณากรอกเหตุผลที่ไม่อนุมัติ!",
-                    },
-                  ]}>
-                  <TextArea
-                    className="col-lg-12"
-                    rows={3}
-                    placeholder="กรอกเหตุผล/เหตุหมายเพิ่มเติม"
-                    autoComplete="off"
-                    onChange={handleOnChangeReason}
-                    defaultValue={data.reason}
-                  />
-                </Form.Item>
-              </div>
-            </div>
-          )}
+          <div className="form-group col-lg-12" style={{ marginTop: 16 }}>
+            <label>หมายเหตุ</label>
+            <Form.Item>
+              <TextArea
+                value={data.comment}
+                onChange={(e) => {
+                  setData((prev) => ({
+                    ...prev,
+                    comment: e.target.value,
+                  }));
+                }}
+              />
+            </Form.Item>
+          </div>
         </Form>
       </CardContainer>
     </div>
@@ -943,7 +940,8 @@ const EditFarmer = () => {
             borderRadius: "12px 12px 0px 0px",
             padding: "10px 10px 10px 10px",
           }}
-          className="d-flex justify-content-between">
+          className="d-flex justify-content-between"
+        >
           <h4 className="pt-2 ps-3" style={{ color: "white" }}>
             แปลงเกษตร
           </h4>
@@ -955,7 +953,8 @@ const EditFarmer = () => {
               border: "none",
               borderRadius: "5px",
             }}
-            onClick={() => setShowAddModal((prev) => !prev)}>
+            onClick={() => setShowAddModal((prev) => !prev)}
+          >
             เพิ่มแปลง
           </Button>
         </div>
@@ -971,11 +970,11 @@ const EditFarmer = () => {
                         overflow: "hidden",
                         whiteSpace: "nowrap",
                         marginBottom: 0,
-                      }}>
+                      }}
+                    >
                       {item.plotName}
                     </p>
-                    <p
-                      style={{ fontSize: "12px", color: color.Grey }}>
+                    <p style={{ fontSize: "12px", color: color.Grey }}>
                       {item.plantName}
                     </p>
                   </div>
@@ -985,23 +984,13 @@ const EditFarmer = () => {
                   <div className="col-lg-3">
                     <span
                       style={{
-                        color:
-                          STATUS_FARMERPLOT_COLOR_MAPPING[
-                            item.status
-                          ],
-                      }}>
+                        color: STATUS_FARMERPLOT_COLOR_MAPPING[item.status],
+                      }}
+                    >
                       <Badge
-                        color={
-                          STATUS_FARMERPLOT_COLOR_MAPPING[
-                            item.status
-                          ]
-                        }
+                        color={STATUS_FARMERPLOT_COLOR_MAPPING[item.status]}
                       />
-                      {
-                        STATUS_NORMAL_MAPPING[
-                          item.status
-                        ]
-                      }
+                      {STATUS_NORMAL_MAPPING[item.status]}
                     </span>
                   </div>
                   <div className="col-lg-3 d-flex justify-content-between">
@@ -1024,9 +1013,7 @@ const EditFarmer = () => {
               ))}
             </div>
           ) : (
-            <div
-              className="container text-center"
-              style={{ padding: "80px" }}>
+            <div className="container text-center" style={{ padding: "80px" }}>
               <img src={emptyData} alt="" />
               <h5>ยังไม่มีแปลงเกษตร</h5>
             </div>