+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## [Version - 0]
+## improve [v0.20.18] 2023-01-26
+- added "REJECTED" status in farmer menu
 ## Fix [v0.20.17] 2023-01-21
 - add farmer plot status fix address to null safety and drafting coupon
 ## Fix [v0.20.16] 2022-12-22
 - image cant remove in EditDroner
 ## Fix [v0.20.15] 2022-12-22
 - change to enter search fired api
-
 ## Fix [v0.20.14] 2022-12-22
 - fixed edit drone image bug
 - change search to onChange fired api